updated empId Type by phase1

+++ src/main/java/com/example/Automate/dto/Employee.java
@@ -2,7 +2,7 @@ package com.example.Automate.dto;
 
 public class Employee {
 
-	private Long Id;
+	private Integer Id;
 	private String name;
 	private Long salary;
 }
